docker-compose provider for kubectl port-forward
Implements a docker-compose provider that "provisions" port forwarding
to resources running on a kubernetes cluster.

The `up` command will spawn a daemon process that manages
`kubectl port-forward` in a subprocess.

`down` will kill that daemon process.

This initial implementation is basic:
- relies on `kubectl` as subshell
- does not handle signals

It basically works, but is not "production grade" yet.